requirements.txt
asgiref==3.2.10
confusable-homoglyphs==3.2.0
Django==3.1.1
django-crispy-forms==1.9.2
django-registration==3.1
djangorestframework==3.11.1
mysqlclient==2.0.1
Pillow==7.2.0
pytz==2020.1
sqlparse==0.3.1
